requirements.txt
# 项目依赖包
requests>=2.28.0
schedule>=1.2.0
urllib3>=1.26.0
pyyaml>=6.0
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0